+++ demo/site/src/config.ts
@@ -1,16 +1,4 @@
 /* eslint-disable no-console */
-export enum aspectRatios {
-    "16x9" = "16x9",
-    "4x3" = "4x3",
-    "3x2" = "3x2",
-    "3x1" = "3x1",
-    "2x1" = "2x1",
-    "1x1" = "1x1",
-    "1x2" = "1x2",
-    "2x3" = "2x3",
-    "3x4" = "3x4",
-    "9x16" = "9x16",
-}
 
 export let domain = "";