+++ packages/api/cms-api/src/builds/builds.service.ts
@@ -8,11 +8,12 @@ import { format } from "date-fns";
 import { CurrentUserInterface } from "../auth/current-user/current-user";
 import { ContentScopeService } from "../content-scope/content-scope.service";
 import { JobStatus } from "../kubernetes/job-status.enum";
+import { INSTANCE_LABEL, PARENT_CRON_JOB_LABEL } from "../kubernetes/kubernetes.constants";
 import { KubernetesService } from "../kubernetes/kubernetes.service";
 import { BuildTemplatesService } from "./build-templates.service";
-import { BUILDER_LABEL, INSTANCE_LABEL, PARENT_CRON_JOB_LABEL, TRIGGER_ANNOTATION } from "./builds.constants";
+import { BUILDER_LABEL, TRIGGER_ANNOTATION } from "./builds.constants";
 import { AutoBuildStatus } from "./dto/auto-build-status.object";
-import { BuildObject } from "./dto/build.object";
+import { Build } from "./dto/build.object";
 import { ChangesSinceLastBuild } from "./entities/changes-since-last-build.entity";
 
 const JOB_HISTORY_LIMIT = 20;
@@ -86,7 +87,7 @@ export class BuildsService {
         return this.createBuilds(trigger, builderCronJobs);
     }
 
-    async getBuilds(user: CurrentUserInterface, options?: { limit?: number | undefined }): Promise<BuildObject[]> {
+    async getBuilds(user: CurrentUserInterface, options?: { limit?: number | undefined }): Promise<Build[]> {
         if (this.kubernetesService.localMode) {
             throw Error("Not available in local mode!");
         }