+++ packages/api/cms-api/src/dam/images/images.controller.ts
@@ -107,7 +107,7 @@ export class ImagesController {
     }
 
     private checkCdnOrigin(incomingCdnOriginHeader: string): void {
-        if (this.config.cdnEnabled) {
+        if (this.config.cdnEnabled && !this.config.disableCdnOriginHeaderCheck) {
             if (incomingCdnOriginHeader !== this.config.cdnOriginHeader) {
                 throw new ForbiddenException();
             }