+++ packages/api/cms-api/src/page-tree/blocks/internal-link.block.ts
@@ -9,7 +9,7 @@ import {
     inputToData,
     TransformResponse,
 } from "@comet/blocks-api";
-import { IsOptional, IsUUID } from "class-validator";
+import { IsOptional, IsString, IsUUID } from "class-validator";
 
 import { PageTreeReadApi } from "../page-tree-read-api";
 import { PageExists } from "../validators/page-exists.validator";
@@ -19,11 +19,16 @@ interface InternalLinkBlockTransformResponse extends TransformResponse {
         id: string;
         name: string;
         path: string;
+        documentType: string;
     } | null;
+    targetPageAnchor?: string;
 }
 
 class InternalLinkBlockData extends BlockData {
     targetPageId?: string;
+
+    @BlockField({ nullable: true })
+    targetPageAnchor?: string;
 
     async transformToPlain({ pageTreeReadApi }: { pageTreeReadApi: PageTreeReadApi }): Promise<InternalLinkBlockTransformResponse> {
         if (pageTreeReadApi === undefined) {
@@ -49,7 +54,9 @@ class InternalLinkBlockData extends BlockData {
                 id: node.id,
                 name: node.name,
                 path: await pageTreeReadApi.nodePath(node),
+                documentType: node.documentType,
             },
+            targetPageAnchor: this.targetPageAnchor,
         };
     }
 }
@@ -60,6 +67,11 @@ class InternalLinkBlockInput extends BlockInput {
     @IsUUID()
     @PageExists()
     targetPageId?: string;
+
+    @BlockField({ nullable: true })
+    @IsOptional()
+    @IsString()
+    targetPageAnchor?: string;
 
     transformToBlockData(): InternalLinkBlockData {
         return inputToData(InternalLinkBlockData, this);
@@ -88,6 +100,11 @@ class Meta extends AnnotationBlockMeta {
                         },
                         {
                             name: "path",
+                            kind: BlockMetaFieldKind.String,
+                            nullable: false,
+                        },
+                        {
+                            name: "documentType",
                             kind: BlockMetaFieldKind.String,
                             nullable: false,
                         },