+++ packages/api/cms-api/src/page-tree/page-tree.service.ts
@@ -188,16 +188,16 @@ export class PageTreeService {
             throw new CometValidationException("Reserved path");
         }
 
+        let newSlug;
         const nodeWithSamePath = await this.nodeWithSamePath(requestedPath, existingNode.scope);
         if (nodeWithSamePath && nodeWithSamePath.id !== existingNode.id) {
-            // @TODO: Or handle gracefully by incrementing the slug ?
-            throw new Error(`Path collision. The path ${requestedPath} is already used`);
+            newSlug = await this.findNextAvailableSlug(existingNode.slug, input.parentId, existingNode.scope);
         }
 
         const parentId = input.parentId;
 
         if (input.pos !== existingNode.pos || input.parentId !== existingNode.parentId) {
-            await this.pageTreeRepository.persistAndFlush(existingNode.assign({ parentId, pos: input.pos }));
+            await this.pageTreeRepository.persistAndFlush(existingNode.assign({ parentId, pos: input.pos, slug: newSlug ?? existingNode.slug }));
 
             const qb = this.pageTreeRepository
                 .createQueryBuilder()
@@ -221,6 +221,24 @@ export class PageTreeService {
         }
 
         return readApi.getNodeOrFail(existingNode.id);
+    }
+
+    async updateNodeSlug(id: string, slug: string): Promise<PageTreeNodeInterface> {
+        const pageTreeReadApi = this.createReadApi({
+            visibility: "all",
+        });
+
+        const node = await pageTreeReadApi.getNodeOrFail(id);
+
+        const requestedPath = await this.pathForParentAndSlug(node.parentId, slug);
+        const nodeWithSamePath = await this.nodeWithSamePath(requestedPath, node.scope);
+        if (nodeWithSamePath && nodeWithSamePath.id !== node.id) {
+            throw new Error("Requested slug is already taken");
+        }
+
+        await this.pageTreeRepository.persistAndFlush(node.assign({ slug: slug }));
+
+        return pageTreeReadApi.getNodeOrFail(id);
     }
 
     async updateCategory(id: string, category: PageTreeNodeCategory): Promise<void> {
@@ -373,5 +391,19 @@ export class PageTreeService {
         const parentPath = existingNodePath.split("/").slice(0, -1).join("/");
         const newPath = `${parentPath || ""}/${slug}`;
         return newPath;
+    }
+
+    private async findNextAvailableSlug(slug: string, parentId: string | null = null, scope?: ScopeInterface): Promise<string> {
+        let counter = 1;
+        let newSlug, newPath;
+
+        do {
+            newSlug = `${slug}-${counter}`;
+            newPath = await this.pathForParentAndSlug(parentId, newSlug);
+
+            counter++;
+        } while (await this.nodeWithSamePath(newPath, scope));
+
+        return newSlug;
     }
 }